Notice tainted installs and update help/doc
If you had local, uncommitted changes in a repository, using the build
command simply recorded the HEAD's hash as the installed version, which
did not reflect what had actually been installed. Add -tainted to the
installed version when running build and recognize it when updating.

Also greatly simplify the documentation comments and help message.